First cut at assigning classes to data and sandboxes. There are currently two annotations: __soaap_classify(L) that allows the assigning of a class to an item of data (currently struct fields) and __soaap_clearance(L) for specifying that a sandboxed method is allowed to access data in class L.
SOAAP picks up these annotations and performs a taint analysis to see where loads of sensitive data propagate to. Every sandboxed load and store instruction is then checked to see if sensitive data is being accessed for which the sandbox does not have clearance for.

Although I use the terms classify and clearance, what I'm trying to introduce is the notion of classes of sandboxes and data and being able to specify (i) membership of this class and (ii) being able to access members of the class. Currently classes are non-hierarchical.