Sandbox instances now store delegated caps.
Previously, the capabilities delegated to a sandbox were found and used during CapabilityAnalysis. However, VulnerabilityAnalysis also needs this information to determine what caps are available to an attacker that exploits a vulnerability in a sandbox. Now, a Sandbox instance obtains all caps that are delegated to it during instance construction.