Updated inconsistent global write analysis.
This analysis now takes into account the __soaap_create_persistent_sandbox("name")
annotations, so that writes to a shared global variable made after the create will
be flagged as not being visible to the sandbox(es) in question. The author could
decide to explicitly sync the writes or defer sandbox creation.

The analysis calculates interprocedurally for each program instruction, the set of
sandbox creation annotations that may have preceded it. A more appropriate name
for this analysis is "reaching sandbox creations".

Also a bug-fix: Correctly identifying whether a global-var/function is extern requires
checking isDeclaration() rather than hasExternalLinkage().