Added support for __soaap_fp on struct members.
Also fixed bug wherein annotated function pointer targets were being cached (in the callToCallees map) but not being added to the callgraph. As a result, because a Sandbox uses the callgraph to find sandboxed functions rather than the cache, these targets were not showing up in the list of sandboxed functions.