Added -soaap-list-fp-calls to list f-pointer calls
This argument to the SOAAP analysis will list all function pointer calls in the LLVM IR and also give a total number at the end.

Also reduced the amount of debug information in LLVM IR by using -gline-tables-only instead of -g. The former only gives us line number information, which is all that we require. However, if we need richer type information then we may have to resort back to -g.