Static call graph building for C++ programs.
We now use LLVM's debugging metadata information to build the class hierarchy for C++ programs and to build a conservative static callgraph using this hierarchy. In particular, for a call x.foo(); the callgraph will contain a edge to all implements of foo() in the class hierarchy starting from x's static type.