Software engineering at work :-)
Turn FPTargetsAnalysis into an abstract class and have our analysis for tracking
annotated function-pointers as a subclass (FPAnnotatedTargetsAnalysis). This will
make way for another analysis for automatically tracking function-pointer assignments
(and thus inferring the targets the function-pointer calls) that reuses a lot of the
code in FPTargetsAnalysis. The end result is that this will hopefully make applying
SOAAP to wireshark easier.