Added support for non-func sandbox code regions.
Previously it was only possible to annotate functions as running in a sandbox.
I've now added a couple of annotations for marking a sandboxed code region as
follows:

__soaap_sandboxed_region_start("renderer")
...
...
...
__soaap_sandboxed_region_end("renderer")

It looks like a latex environment :).

These region annotations should be used in a structured way, i.e. in the same
scope.

Still more work is needed here to also specify clearances and performance
information for code regions.