Fix SOAAP in line with changes to LLVM API semantics.
Dereferencing a Use object would previously give the User but now gives the Value that is being used. This change meant that CapabilityAnalysis and PrivilegedCallAnalysis broke. We now use the iterator range function Value::users() to be able to use range-based for loops to iterate over calls of a function.