Bugfix: Only resize llvm bitvectors if they are too small.
The resize(N) function on llvm BitVectors will either grow or shrink the BitVector depending on whethere N is smaller or greater than the current number of bits it contains. SOAAP uses resize() to ensure that a particular bit index exists before setting it, but a check for if idx >= number of bits was not being made leading to shrinking in some scenarios.

This led to a propagation bug for wireshark whereby some of the system calls that were annotated as being allowed were lost and thus did not propagate through the program.