Use unordered_set instead of set.
set<string>::insert(...) is triggering a bizarre segfault in libc++ r208080 and current HEAD, when running a debug build of SOAAP. David Chisnall has also helped debug this issue and we can't quite work out what exactly is going on. unordered_set doesn't cause a segfault so using that instead.