Create the persistent sandbox.
It seems that we need to annotate the place where we create the sandbox, not just the function that represents the sandbox.

This still doesn't fix the test (SOAAP complains about both the read and the write but we've said that reads are ok), but it crystalizes exactly what's wrong with it.