Combine CapabilityAnalysis and CapabilitySysCallsAnalysis. CapabilityAnalysis and CapabilitySysCallsAnalysis pretty much perform the same analysis: Determine if system calls are being performed on a file descriptor, whereby the necessary file descriptor rights have not been annotated. However, they each use two different sets of annotations:
__soaap_fd_permit(fd, syscalls...)

and

__soaap_limit_fd_syscalls(fd, syscalls...)
__soaap_limit_fd_key_syscalls(fdkey, syscalls...)

Having them as two separate analyses results in each analysis only considering its respective annotation to mean "annotated with rights". For example, CapabilityAnalysis looks at __soaap_fd_permit and ignores the __soaap_limit_* annotations. This results in SOAAP producing warnings about rights not being possessed, but which have actually been specified using the other annotation.

To fix this, we merge the two analyses into the CapabilitySysCallsAnalysis class. In the future we may normalise things so that we have only one set of annotations to express the allowable system calls on a file descriptor. However, for the time being we will keep both types of annotations.