Add support for sandbox-platform modelling.
Add command-line option --soaap-sandbox-platform=<annotated|capsicum|seccomp>, which when passed to the SOAAP command-line tool models the (currently only) system call semantics of the corresponding platform. For example, if capsicum is specified then system calls such as open are disallowed in sandboxes while read, write, etc, require explicit rights on the file descriptor to have been annotated using __soaap_limit_fd_syscalls or __soaap_limit_fdkey_syscalls.

To support this, a SandboxPlatform abstract class has been added as the base of sandbox platform descriptions. Currently, there are concrete subclasses for Capsicum and Seccomp. SysCallAnalysis and CapabilitySysCallAnalysis have been updated to check with the platform if it has been specified.

The 'annotated' option uses the __soaap_limit_syscalls annotation to determine which system calls may be called in a sandbox and __soaap_limit_fd_syscalls annotation to specify restrictions (if any) on file descriptors. SOAAP will use these annotations when determining whether a system call is allowed or not.

The default value of this command-line option is capsicum.