Make soaap_declare_callgates_helper function extern to prevent inlining.
SOAAP inserts annotated callgate information as a call to a function called soaap_declare_callgates_helper, with the callgate functions as arguments. This function has an empty body and so is inlined by clang at -O1 and above. Given that it is called from an unused function called soaap_declare_callgates, we can instead make it extern to prevent inlining.