Fix invalid call cycles.
Once upon a time, we couldn't assume that all nodes in the call graph had location information, so we treated the `old` node differently from the `callee` node. When we got location information for all nodes, however, we didn't update this distinction, so sometimes we ended up using a stale `callee` variable. Among other things, this resulted in things apparently calling `main`!