Finish support for context-sensitive function-pointer inference and callgraph.
SOAAP now tracks function-pointer flows in a context-sensitive manner and the callgraph can now be queried for the callees for the current context (which is either "privileged or the current sandbox).

Context-sensitive is turned on by default. To turn it off, use the --soaap-context-insens flag.

Nothing special is needed to use the new function-pointer inference except the usual --soaap-infer-fp-targets.

This change has required quite a bit of restructuring (and cleanup) of SOAAP.