Convert libxo module to pull directly from GitHub.
Rather than including the libxo sources as a SOAAP submodule, teach
SOAAP's CMake infrastructure how to download libxo with tag 0.1.6.

On a separate note, we might want to update to a more recent libxo at some
point. At the very least, it would be convenient to use exactly the
same version as FreeBSD.