Add --soaap-warn-modules and --soaap-nowarn-modules command-line options.
The first will cause SOAAP to only output warnings for a module in the
comma-separate list of modules and no other. The second will cause SOAAP to not
output warnings for the list of modules specified.

Note that an error will be produced if both command-line options are used
together.

Names of modules are as they are inserted by llvm-link.