Fix CMake runs failing
CMake seems to have problems with setting CMAKE_C_COMPILER in the
CMakeLists.txt. It will rerun CMake but clear the cache. In that case
LLVM_BUILD_DIR will no longer be set so we would get an error.
To work around this issue we can infer the path from CMAKE_C_COMPILER
since that should be ${LLVM_BUILD_DIR}/bin/clang. To check whether it
is really a clang build directory and not /usr/bin/clang we can check
whether COMPILER_DIR/../LLVMBuild.cmake exists since that will only be
the in a build directory and not in /usr