Add RAII wrappers for XO::[open/close]_[container/instance/list]()
They make sure that the element is closed on scope exit or when close()
is called. This make sure that there is no name mismatch between the
open_*() and close_*() call. Additionally XO::Instance can only be
constructed from a const reference to XO::List which ensures that these
names also match.