Generate IR: support ar append mode
This is quite problematic since it might result in duplicated symbols.
This is not a problem for ar since it discards them, but causes llvm-link
to fail.
We handle this case by moving the target file to a temporary directory
and appending that moved file to the command line. This will cause
errors if the existing .a.bc file already contains the files that will
be linked. As a workaround you can run
`find . -name "*.a.bc" -print0 | xargs -0 rm`