Tolerate already-applied libxo patch.
`patch` returns a non-zero exit code if we tell it to apply forward patches
only and the only hunk has already been applied. So, if patch "fails",
let's use `|| true` to keep the build going. If the patch caused serious
problems then that'll be reflected in the next build step.